Rename SettingsLayout to BuildLayout
- remove FileSystemLayout

Issue: #16506
