+++ platforms/core-configuration/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/SettingsDelegate.kt
@@ -32,7 +32,7 @@ import org.gradle.api.plugins.PluginManager
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.api.toolchain.management.ToolchainManagement
 import org.gradle.api.cache.CacheConfigurations
-import org.gradle.api.file.SettingsLayout
+import org.gradle.api.file.BuildLayout
 import org.gradle.caching.configuration.BuildCacheConfiguration
 import org.gradle.internal.deprecation.DeprecationLogger
 import org.gradle.plugin.management.PluginManagementSpec
@@ -167,6 +167,6 @@ abstract class SettingsDelegate : Settings {
 
     override fun caches(cacheConfigurations: Action<in CacheConfigurations>) = delegate.caches(cacheConfigurations)
 
-    override fun getLayout(): SettingsLayout =
+    override fun getLayout(): BuildLayout =
         delegate.layout
 }
