+++ subprojects/core-api/src/main/java/org/gradle/api/file/BuildLayout.java
@@ -18,17 +18,11 @@ package org.gradle.api.file;
 
 import org.gradle.api.Incubating;
 import org.gradle.api.initialization.Settings;
-import org.gradle.api.provider.Provider;
 import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
-import java.io.File;
-
 /**
- * Provides access to important locations for a {@link Settings} instance.
- * <p>
- * File system locations (or providers thereof) based on relative paths will be
- * resolved against this layout's reference location, as defined by {@link #getSettingsDirectory()}.
+ * Provides access to important locations for a Gradle build.
  * <p>
  * An instance of this type can be injected into a plugin or other object by
  * annotating a public constructor or method with {@code javax.inject.Inject}.
@@ -40,7 +34,7 @@ import java.io.File;
  */
 @Incubating
 @ServiceScope(Scopes.Build.class)
-public interface SettingsLayout extends FileSystemLayout {
+public interface BuildLayout {
     /**
      * Returns the settings directory.
      * <p>
@@ -58,25 +52,4 @@ public interface SettingsLayout extends FileSystemLayout {
      * @see Settings#getRootDir()
      */
     Directory getRootDirectory();
-
-
-    /**
-     * {@inheritDoc}
-     *
-     * <p>
-     * File system locations based on relative paths will be
-     * resolved against this layout's reference location, as defined by {@link #getSettingsDirectory()}.
-     */
-    @Override
-    Provider<RegularFile> file(Provider<File> file);
-
-    /**
-     * {@inheritDoc}
-     *
-     * <p>
-     * File system locations based on relative paths will be
-     * resolved against this layout's reference location, as defined by {@link #getSettingsDirectory()}.
-     */
-    @Override
-    Provider<Directory> dir(Provider<File> file);
 }
