+++ subprojects/core-api/src/main/java/org/gradle/api/file/ProjectLayout.java
@@ -25,9 +25,6 @@ import java.io.File;
 
 /**
  * Provides access to several important locations for a project.
- * <p>
- * File system locations (or providers thereof) based on relative paths will be
- * resolved against this layout's reference location, as defined by {@link #getProjectDirectory()}.
  *
  * <p>An instance of this type can be injected into a task, plugin or other object by annotating a public constructor or method with {@code javax.inject.Inject}. It is also available via {@link org.gradle.api.Project#getLayout()}.
  *
@@ -35,7 +32,7 @@ import java.io.File;
  * @since 4.1
  */
 @ServiceScope(Scopes.Project.class)
-public interface ProjectLayout extends FileSystemLayout {
+public interface ProjectLayout {
     /**
      * Returns the project directory.
      */
@@ -47,23 +44,21 @@ public interface ProjectLayout extends FileSystemLayout {
     DirectoryProperty getBuildDirectory();
 
     /**
-     * {@inheritDoc}
+     * Creates a {@link RegularFile} provider whose location is calculated from the given {@link Provider}.
      * <p>
      * File system locations based on relative paths will be
      * resolved against this layout's reference location, as defined by {@link #getProjectDirectory()}.
      */
-    @Override
     Provider<RegularFile> file(Provider<File> file);
 
     /**
-     * {@inheritDoc}
+     * Creates a {@link Directory} provider whose location is calculated from the given {@link Provider}.
      * <p>
      * File system locations based on relative paths will be
      * resolved against this layout's reference location, as defined by {@link #getProjectDirectory()}.
      *
      * @since 6.0
      */
-    @Override
     Provider<Directory> dir(Provider<File> file);
 
     /**
