+++ subprojects/core-api/src/main/java/org/gradle/api/initialization/Settings.java
@@ -20,7 +20,7 @@ import org.gradle.StartParameter;
 import org.gradle.api.Action;
 import org.gradle.api.Incubating;
 import org.gradle.api.UnknownProjectException;
-import org.gradle.api.file.SettingsLayout;
+import org.gradle.api.file.BuildLayout;
 import org.gradle.api.initialization.dsl.ScriptHandler;
 import org.gradle.api.initialization.resolve.DependencyResolutionManagement;
 import org.gradle.api.invocation.Gradle;
@@ -198,12 +198,12 @@ public interface Settings extends PluginAware, ExtensionAware {
     Settings getSettings();
 
     /**
-     * Provides access to various important directories for this settings object.
+     * Provides access to important locations for a Gradle build.
      *
      * @since 8.5
      */
     @Incubating
-    SettingsLayout getLayout();
+    BuildLayout getLayout();
 
     /**
      * Returns the build script handler for settings. You can use this handler to query details about the build
