+++ subprojects/core/src/integTest/groovy/org/gradle/api/file/ProjectLayoutIntegrationTest.groovy
@@ -16,7 +16,6 @@
 
 package org.gradle.api.file
 
-
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 
 class ProjectLayoutIntegrationTest extends AbstractIntegrationSpec {
