+++ subprojects/docs/src/docs/dsl/dsl.xml
@@ -148,7 +148,7 @@
                 <td>org.gradle.api.file.ProjectLayout</td>
             </tr>
             <tr>
-                <td>org.gradle.api.file.SettingsLayout</td>
+                <td>org.gradle.api.file.BuildLayout</td>
             </tr>
             <tr>
                 <td>org.gradle.api.Script</td>
